backend/requirements.txt
Django==5.1.7
psycopg2-binary==2.9.10
gunicorn==23.0.0
python-dotenv==1.1.0
djangorestframework==3.16.0